// types/location/country/continent.type.ts
// ======================================================
// CONTINENT
// ======================================================

import type { Country } from "./country";

export interface Continent {
  id: string;

  // ======================================================
  // BASIC
  // ======================================================

  name: string;

  nativeName?: string | null;

  code: string;

  slug: string;

  description?: string | null;

  // ======================================================
  // MEDIA
  // ======================================================

  thumbnail?: string | null;

  coverImage?: string | null;

  // ======================================================
  // DISPLAY
  // ======================================================

  sortOrder: number;

  // ======================================================
  // STATUS
  // ======================================================

  active: boolean;

  // ======================================================
  // RELATIONS
  // ======================================================

  countries: Country[];

  // ======================================================
  // TIMESTAMP
  // ======================================================

  createdAt: Date;

  updatedAt: Date;
}